seller table added

+++ database/migrations/2019_06_15_042320_create_currencies_table.php
@@ -17,7 +17,7 @@ class CreateCurrenciesTable extends Migration
             $table->increments('id')->unsigned();
             $table->string('language_id')->nullable();
             # name fetch from dictionary
-            $table->float('ratio', 18, 9);
+            $table->decimal('ratio', 18, 9);
             $table->foreign('language_id')->references('id')->on('languages');
         });
     }
